API/requirements.txt
fastapi
uvicorn
python-multipart
mysql-connector-python
cachetools
requests
python-dotenv
pyjwt
passlib